% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge_editor.R
\name{forgeEditorOutput}
\alias{forgeEditorOutput}
\title{Shiny output binding for forge_editor}
\usage{
forgeEditorOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output slot ID.}

\item{width, height}{Widget dimensions.}
}
\description{
Shiny output binding for forge_editor
}
