% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge_editor.R
\name{renderForgeEditor}
\alias{renderForgeEditor}
\title{Shiny render binding for forge_editor}
\usage{
renderForgeEditor(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A `forge_editor()` call.}

\item{env}{Environment to evaluate `expr` in.}

\item{quoted}{Whether `expr` is already quoted.}
}
\description{
Shiny render binding for forge_editor
}
